fastapi-shapefile-microservice/requirements.txt
azure-storage-blob
geopandas
psycopg2
sqlalchemy
fastapi 
uvicorn